[fixed] deleting active resource spawns from db
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@2120 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/resource/ResourceSpawnImplementation.cpp
@@ -200,7 +200,7 @@ ResourceContainer* ResourceSpawnImplementation::createResource(int units) {
 }
 
 void ResourceSpawnImplementation::decreaseContainerReferenceCount() {
-	if (--containerReferenceCount < 1) {
+	if (--containerReferenceCount < 1 && !inShift()) {
 		destroyObjectFromDatabase(true);
 
 		dbDestroyed = true;
